+++ R/fieldbook.R
@@ -1,0 +1,19 @@
+#' Fieldbook experimental design
+#'
+#' The dataset were obtained using inti package. The data set is a randomize
+#' complete block design (RCBD) with three replications.
+#'
+#' @format A data frame with 24 rows and 5 variables:
+#'
+#'   \describe{
+#'
+#'   \item{barcode}{barcode for each experimental unit} 
+#'   \item{plots}{Plot number} 
+#'   \item{block}{Blocks (3): number of replication in the design}
+#'   \item{condition}{Factor with two levels: irrigated and drought}
+#'   \item{genotypes}{Factor with four levels: choclito, salcedo, pandela, puno}
+#'
+#'   }
+#'   
+
+"fieldbook"