+++ R/include_text.R
@@ -16,27 +16,6 @@
 #'
 #' @export
 #'
-#' @examples
-#'
-#' \donttest{
-#'
-#' library(gsheet)
-#'
-#' url <- paste0("https://docs.google.com/spreadsheets/d/"
-#'        , "1q0EZmZBt52ca-0VbididjJy2jXTwf06laJpzvkQJWvc/edit#gid=107939497")
-#' fb <- gsheet2tbl(url)
-#'
-#' label <- label_layout(data = fb
-#'                , size = c(10, 2.5)
-#'                ) %>%
-#'          include_image(value = "https://inkaverse.com/reference/figures/logo.png"
-#'                , size = c(2,2)
-#'                , position = "1.2*1.25"
-#'                ) %>%
-#'          include_text(value = "plots", position = "5*1.25") %>% 
-#'          label_print()
-#'
-#' }
 #'
 
 include_text <- function(label