+++ NEWS.md
@@ -3,6 +3,8 @@
 - Update vignettes
 - Not load any font by default (i.e Permanent Marker)
 - `label_layout()` use tibble as default
+- Include a dataset: `fieldbook`
+- Update examples in functions
 
 # huito 0.1.2